+++ .changeset/small-dryers-join.md
@@ -1,0 +1,7 @@
+---
+'@leafygreen-ui/checkbox': major
+---
+
+Updates checkbox for visual brand refresh
+ - Removes `bold` prop
+ - Adds `description` prop
