+++ packages/checkbox/package.json
@@ -21,7 +21,8 @@
   },
   "dependencies": {
     "@leafygreen-ui/hooks": "^7.0.0",
-    "@leafygreen-ui/lib": "^9.0.0"
+    "@leafygreen-ui/lib": "^9.0.0",
+    "@leafygreen-ui/typography": "^9.1.0"
   },
   "devDependencies": {
     "@leafygreen-ui/palette": "^3.2.2"
