+++ packages/checkbox/tsconfig.json
@@ -18,6 +18,9 @@
     },
     {
       "path": "../hooks"
+    },
+    {
+      "path": "../typography"
     }
   ]
 }
